Match wispbaddie_update signed int->f32 conversion
(f32)(int) not (f32)(u32)(int) on randomGetRange forces the signed-magic
xoris;lfd path (recipe #10). wispbaddie_update 99.58 -> 100%.

(cherry picked from commit 4c5130f44a6e91846520288d7e4773d4514bcb56)

+++ src/main/dll/dll_00E1_wispbaddie.c
@@ -357,7 +357,7 @@ void wispbaddie_update(int obj)
         if (state->cryTimer < lbl_803E2714)
         {
             Sfx_PlayFromObject(obj, 0x23d);
-            state->cryTimer = (f32)(u32)(int)randomGetRange(0x3c, 0x78);
+            state->cryTimer = (f32)(int)randomGetRange(0x3c, 0x78);
         }
         state->particleId = 0x338;
     }
